Refactor CMakeLists.txt
- Fix include path
- C++Now 2017: Daniel Pfeifer "Effective CMake"
  * https://www.youtube.com/watch?v=bsXLMQ6WgIk
