+++ src/Plugin/src/Export/CsvExport/CMakeLists.txt
@@ -23,8 +23,7 @@ target_sources(${LIBRARY_NAME}
         src/FlightRadar24CsvWriter.h src/FlightRadar24CsvWriter.cpp
         src/CsvExportPlugin.json
 )
-
-target_include_directories(${LIBRARY_NAME} PUBLIC
-    "${CMAKE_CURRENT_BINARY_DIR}"
+target_link_libraries(${LIBRARY_NAME}
+    PRIVATE
+        ${LIBRARY_DEPENDENCIES}
 )
-target_link_libraries(${LIBRARY_NAME} PRIVATE ${LIBRARY_DEPENDENCIES})
