+++ src/Plugin/src/Export/GpxExport/CMakeLists.txt
@@ -20,8 +20,7 @@ target_sources(${LIBRARY_NAME}
         src/GpxExportOptionWidget.h src/GpxExportOptionWidget.cpp src/GpxExportOptionWidget.ui
         src/GpxExportPlugin.json
 )
-
-target_include_directories(${LIBRARY_NAME} PUBLIC
-    "${CMAKE_CURRENT_BINARY_DIR}"
+target_link_libraries(${LIBRARY_NAME}
+    PRIVATE
+        ${LIBRARY_DEPENDENCIES}
 )
-target_link_libraries(${LIBRARY_NAME} PRIVATE ${LIBRARY_DEPENDENCIES})
