+++ src/Plugin/src/Export/IgcExport/src/IgcExportSettings.h
@@ -31,7 +31,7 @@
 #include <QString>
 
 #include "../../../../../Kernel/src/Settings.h"
-#include "../../Plugin/src/ExportPluginBaseSettings.h"
+#include "../../../ExportPluginBaseSettings.h"
 
 class IgcExportSettingsPrivate;
