+++ src/Plugin/src/Export/JsonExport/src/JsonExportSettings.h
@@ -30,7 +30,7 @@
 #include <QObject>
 
 #include "../../../../../Kernel/src/Settings.h"
-#include "../../Plugin/src/ExportPluginBaseSettings.h"
+#include "../../../ExportPluginBaseSettings.h"
 
 class JsonExportSettingsPrivate;
