+++ src/Plugin/src/Export/KmlExport/src/KmlExportSettings.h
@@ -31,7 +31,7 @@
 #include <QColor>
 
 #include "../../../../../Kernel/src/Settings.h"
-#include "../../Plugin/src/ExportPluginBaseSettings.h"
+#include "../../../ExportPluginBaseSettings.h"
 
 class KmlExportSettingsPrivate;
