+++ src/Plugin/src/Import/CsvImport/CMakeLists.txt
@@ -30,5 +30,7 @@ target_sources(${LIBRARY_NAME}
         src/CsvImportPlugin.json
         ${DOCUMENTATION}
 )
-
-target_link_libraries(${LIBRARY_NAME} PRIVATE ${LIBRARY_DEPENDENCIES})
+target_link_libraries(${LIBRARY_NAME}
+    PRIVATE
+        ${LIBRARY_DEPENDENCIES}
+)
