+++ src/Plugin/src/Import/IgcImport/src/IgcImportPlugin.cpp
@@ -35,6 +35,8 @@
 #include <QDateTime>
 #include <QEasingCurve>
 
+#include <GeographicLib/Geoid.hpp>
+
 #include "../../../../../Kernel/src/Unit.h"
 #include "../../../../../Kernel/src/Settings.h"
 #include "../../../../../Kernel/src/SkyMath.h"
@@ -83,6 +85,8 @@ public:
 
     IgcImportSettings pluginSettings;
     QEasingCurve throttleResponseCurve;
+
+    GeographicLib::Geoid egm96 {"egm2008-2_5", QCoreApplication::applicationDirPath().append("/geoids").toStdString()};
 
     static const inline QString FileSuffix {QStringLiteral("igc")};
 };
@@ -144,10 +148,26 @@ bool IgcImportPlugin::importFlight(QFile &file, Flight &flight) noexcept
         for (const IgcParser::Fix &fix : d->igcParser.getFixes()) {
             // Import either GNSS or pressure altitude
             const double altitude = d->pluginSettings.getAltitudeMode() == IgcImportSettings::AltitudeMode::Gnss ? fix.gnssAltitude : fix.pressureAltitude;
-            PositionData positionData {fix.latitude, fix.longitude, altitude};
-            positionData.timestamp = fix.timestamp;
-            positionData.indicatedAltitude = fix.pressureAltitude;
-            position.upsertLast(std::move(positionData));
+
+            // @todo FIX ME It is rather pointless to adjust the reported heights with "undulated values": the reported
+            //              altitude is pressure altitude (not GNSS altitude). Also, altitude while on ground is throughouly reported as 0.
+            try {
+
+              // Convert height above egm96 to height above the ellipsoid
+
+              double geoid_height = d->egm96(fix.latitude, fix.longitude);
+              double height_above_geoid = (altitude + GeographicLib::Geoid::ELLIPSOIDTOGEOID * geoid_height);
+              qDebug("height_above_geoid: %f", height_above_geoid);
+
+              PositionData positionData {fix.latitude, fix.longitude, Convert::metersToFeet(height_above_geoid)};
+              positionData.timestamp = fix.timestamp;
+              positionData.indicatedAltitude = Convert::metersToFeet(fix.pressureAltitude);
+              position.upsertLast(std::move(positionData));
+            }
+            catch (const std::exception& e) {
+              qDebug("Caught exception: %s", e.what());
+
+            }
 
             if (d->igcParser.hasEnvironmentalNoiseLevel()) {
                 const double enl = fix.environmentalNoiseLevel;
