+++ src/Plugin/src/Import/KmlImport/doc/README.md
@@ -19,7 +19,16 @@ Various XML elements may store coordinates (latitude/longitude/altitude), as chi
   * Connects two or more coordinates
   * No associated timestamps
   * Example: KML exported from flightradar24.com
+  
+## Altitude
+
+Both FlightAware and FlightRadar24 (respectively transponders and ADS-B) report pressure altitudes at standard pressure (29.92 In-Hg or 1013.25 hPa). FlightRadar24 also reports all altitudes to be 0 at ground level. Also, ADS-B can only encode the altitude to the nearest 25 feet.
+
+FlightAware reports "absolute" (KML: "above mean sea level") altitude values, which may lead "into the ground" (below runways) or be "above the runway" (depending on the actual pressure measured at the time).
+
+FlightRadar24 on the other hand reports 0 altitude values while the aircraft is on the ground (otherwise pressure altitude, as discussed above). Interestingly while the KML Placemarks have an "absolute" altitude mode (putting them "under the ground as well") the actual track data has (properly) a "clampToGround" KML altitude mode, aligning the visible track with runways and taxiways.
 
 ## Resources
 
 - KML documentation by Google: https://developers.google.com/kml/documentation/
+- FlightAware altitude discussion: https://www.pilotsofamerica.com/community/threads/flightaware-altitude-reporting.129363/
