+++ src/Plugin/src/Import/KmlImport/src/FlightRadar24KmlParser.cpp
@@ -24,6 +24,7 @@
  */
 #include <memory>
 #include <cstdint>
+#include <exception>
 
 #include <QString>
 #include <QStringLiteral>
@@ -31,6 +32,8 @@
 #include <QDateTime>
 #include <QXmlStreamReader>
 #include <QRegularExpression>
+
+#include <GeographicLib/Geoid.hpp>
 
 #include "../../../../../Kernel/src/Convert.h"
 #include "../../../../../Model/src/Flight.h"
@@ -72,6 +75,8 @@ public:
 
     QRegularExpression speedRegExp;
     QRegularExpression headingRegExp;
+
+    GeographicLib::Geoid egm96 {"egm2008-2_5", QCoreApplication::applicationDirPath().append("/geoids").toStdString()};
 
 private:
     static inline const QString SpeedPattern {QStringLiteral("<b>Speed:<\\/b><\\/span> <span>(\\d+) kt<\\/span>")};
@@ -303,7 +308,26 @@ void FlightRadar24KmlParser::parsePoint() noexcept
                     auto &trackItem = d->trackData.back();
                     trackItem.latitude = latitude;
                     trackItem.longitude = longitude;
-                    trackItem.altitude = Convert::metersToFeet(altitude);
+
+                    // @todo FIX ME It is rather pointless to adjust the reported heights with "undulated values": the reported
+                    //              altitude is pressure altitude (not GNSS altitude). Also, altitude while on ground is throughouly reported as 0.
+                    try {
+
+                      // Convert height above egm96 to height above the ellipsoid
+
+                      double geoid_height = d->egm96(latitude, longitude);
+                      double height_above_geoid = (altitude + GeographicLib::Geoid::ELLIPSOIDTOGEOID * geoid_height);
+                      qDebug("height_above_geoid: %f", height_above_geoid);
+                      trackItem.altitude = Convert::metersToFeet(height_above_geoid);
+                    }
+                    catch (const std::exception& e) {
+                      qDebug("Caught exception: %s", e.what());
+
+                    }
+
+
+
+
                 }
 
             } else {
