+++ src/UserInterface/CMakeLists.txt
@@ -41,7 +41,13 @@ target_sources(${LIBRARY_NAME}
         src/Widget/WaypointWidget.h src/Widget/WaypointWidget.cpp src/Widget/WaypointWidget.ui
 )
 
-include_directories(UserInterface PRIVATE ../Widget/src/)
-target_link_libraries(${LIBRARY_NAME} PRIVATE ${LIBRARY_DEPENDENCIES})
-set_target_properties(${LIBRARY_NAME} PROPERTIES VERSION ${PROJECT_VERSION})
-set_target_properties(${LIBRARY_NAME} PROPERTIES SOVERSION ${PROJECT_VERSION_MAJOR})
+target_include_directories(${LIBRARY_NAME} PRIVATE ../Widget/src/)
+target_link_libraries(${LIBRARY_NAME}
+    PRIVATE
+        ${LIBRARY_DEPENDENCIES}
+)
+set_target_properties(${LIBRARY_NAME}
+    PROPERTIES
+        VERSION ${PROJECT_VERSION}
+        SOVERSION ${PROJECT_VERSION_MAJOR}
+)
